Começando os botões do jogo
